Remove skins with duplicating class names

+++ install/_functions_evoupgrade.php
@@ -8643,39 +8643,58 @@ function upgrade_b2evo_tables( $upgrade_action = 'evoupgrade' )
 		// Array key is ID of old skin, Array value is ID of new skin with folder suffix "_skin":
 		$duplicating_skins = $DB->get_assoc( $skins_SQL );
 
-		// Get collections with old skin format folder(wihtout suffix "_skin"):
-		$old_skins = array_keys( $duplicating_skins );
-		$skin_fields = array( 'blog_normal_skin_ID', 'blog_mobile_skin_ID', 'blog_tablet_skin_ID' );
-		$blog_SQL = new SQL( 'Get collection with dupplicating skin class names' );
-		$blog_SQL->SELECT( 'blog_ID, blog_normal_skin_ID, blog_mobile_skin_ID, blog_tablet_skin_ID' );
-		$blog_SQL->FROM( 'T_blogs' );
-		foreach( $skin_fields as $skin_field )
+		if( count( $duplicating_skins ) > 0 )
 		{
-			$blog_SQL->WHERE_or( $skin_field.' IN ( '.$DB->quote( $old_skins ).' )' );
-		}
-		$old_skin_colls = $DB->get_results( $blog_SQL );
-
-		foreach( $old_skin_colls as $old_skin_coll )
-		{
-			$coll_update_fields = array();
+			// Get collections with old skin format folder(wihtout suffix "_skin"):
+			$old_skins = array_keys( $duplicating_skins );
+			$skin_fields = array( 'blog_normal_skin_ID', 'blog_mobile_skin_ID', 'blog_tablet_skin_ID' );
+			$blog_SQL = new SQL( 'Get collection with dupplicating skin class names' );
+			$blog_SQL->SELECT( 'blog_ID, blog_normal_skin_ID, blog_mobile_skin_ID, blog_tablet_skin_ID' );
+			$blog_SQL->FROM( 'T_blogs' );
 			foreach( $skin_fields as $skin_field )
 			{
-				if( isset( $duplicating_skins[ $old_skin_coll->$skin_field ] ) )
-				{	// If the skin is old folder format then use skin ID of new format skin folder:
-					$coll_update_fields[] = $skin_field.' = '.$duplicating_skins[ $old_skin_coll->$skin_field ];
-				}
+				$blog_SQL->WHERE_or( $skin_field.' IN ( '.$DB->quote( $old_skins ).' )' );
 			}
-			// Update collection skin IDs:
-			$DB->query( 'UPDATE T_blogs
-				  SET '.implode( ', ', $coll_update_fields ).'
-				WHERE blog_ID = '.$old_skin_coll->blog_ID );
-		}
+			$old_skin_colls = $DB->get_results( $blog_SQL );
 
-		// Remove duplicating skins with old format folders:
-		$DB->query( 'DELETE FROM T_skins__skin
-			WHERE skin_ID IN ( '.$DB->quote( $old_skins ).' )' );
-		$DB->query( 'DELETE FROM T_skins__container
-			WHERE sco_skin_ID IN ( '.$DB->quote( $old_skins ).' )' );
+			// Update collections:
+			foreach( $old_skin_colls as $old_skin_coll )
+			{
+				$coll_update_fields = array();
+				$coll_settings_update_fields = array();
+				foreach( $skin_fields as $skin_field )
+				{
+					if( isset( $duplicating_skins[ $old_skin_coll->$skin_field ] ) )
+					{	// If the skin is old folder format then use skin ID of new format skin folder:
+						$coll_update_fields[] = $skin_field.' = '.$duplicating_skins[ $old_skin_coll->$skin_field ];
+						$coll_settings_update_fields[] = $skin_field.' = '.$duplicating_skins[ $old_skin_coll->$skin_field ];
+					}
+				}
+				// Update collection skin IDs:
+				$DB->query( 'UPDATE T_blogs
+					  SET '.implode( ', ', $coll_update_fields ).'
+					WHERE blog_ID = '.$old_skin_coll->blog_ID );
+			}
+
+			foreach( $duplicating_skins as $old_skin_ID => $new_skin_ID )
+			{
+				// Update custom skin settings of collections:
+				$DB->query( 'UPDATE T_coll_settings
+					  SET cset_name = CONCAT( '.$DB->quote( 'skin'.$new_skin_ID.'_' ).', SUBSTRING( cset_name, '.( strlen( 'skin'.$old_skin_ID.'_' ) + 1 ).' ) )
+					WHERE cset_name LIKE "skin'.$old_skin_ID.'_%" ' );
+				// Update default skins:
+				$DB->query( 'UPDATE T_settings
+					  SET set_value = '.$new_skin_ID.'
+					WHERE set_name LIKE "%_skin_ID"
+					  AND set_value LIKE '.$old_skin_ID );
+			}
+
+			// Remove duplicating skins with old format folders:
+			$DB->query( 'DELETE FROM T_skins__skin
+				WHERE skin_ID IN ( '.$DB->quote( $old_skins ).' )' );
+			$DB->query( 'DELETE FROM T_skins__container
+				WHERE sco_skin_ID IN ( '.$DB->quote( $old_skins ).' )' );
+		}
 
 		upg_task_end();
 	}
